Nessie: Refactor NessieTableOperations
